+++ Q3/PacletInfo.m
@@ -4,7 +4,7 @@
 
 Paclet[
     Name -> "Q3",
-    Version -> "1.3.1",
+    Version -> "1.3.2",
     MathematicaVersion -> "6+",
     Description -> "Mathematica application to help study quantum information processing, quantum many-body systems, and quantum spin systems.",
     Creator -> "Mahn-Soo Choi (Korea University)",